pyproject.toml
[tool.poetry]
name = "nostradamus"
version = "0.1.0"
description = ""
authors = ["ihsan <ihsan.m@ampcome.com>"]
package-mode = false
[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
python-telegram-bot = "^21.10"
requests = "^2.32.3"
telegramify-markdown = {extras = ["mermaid"], version = "^0.4.2"}
markdown-it-py = "^3.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"

[poetry.group.dev.dependencies]
optional = true

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"